feat: Add wishlist functionality
This commit introduces the ability for users to save products to a wishlist.

The UserProfile type has been updated to include a `wishlist` field.
The SmartShelf component now supports moving items to the shelf and removing them from the wishlist.
The App component has been modified to handle adding and removing products from the wishlist, as well as moving products to the shelf.
The PremiumRoutineBuilder component now includes an option to add products to the wishlist.